reafactor: 계산을 도와주는 Enum을 내부에서 외부로 선언으로 변경
