+++ OOP_Prcatice/java-lotto/src/main/java/lotto/model/CalculatorRanking.java
@@ -1,28 +1,6 @@
 package lotto.model;
 
-enum calculatorHelper {
-    INDEX_ARRAY_HELPER(3),
-    INDEX_SIZE(5),
-
-    NEED_CHECK_NUMBER(5),
-    ALL_RIGHT(6),
-    MIN_RIGHT(3),
-    ALL_RIGHT_INDEX(4),
-    NO_BONUS_INDEX(2),
-    YES_BONUS_INDEX(3),
-    NO_RANK_INDEX(-1),
-    ;
-    private int helpNumber;
-
-    calculatorHelper(int helpNumber) {
-        this.helpNumber = helpNumber;
-    }
-
-    public int getHelpNumber() {
-        return helpNumber;
-    }
-}
-
+import lotto.calculatorHelper;
 
 public class CalculatorRanking {
 
@@ -40,8 +18,8 @@ public class CalculatorRanking {
 
     private void calculateRanking() {
         for (Lotto lotto : lottoGenerator.getLottos()) {
-            int index=calculateDetail(lotto);
-            if(index!=calculatorHelper.NO_RANK_INDEX.getHelpNumber()) {
+            int index = calculateDetail(lotto);
+            if (index != calculatorHelper.NO_RANK_INDEX.getHelpNumber()) {
                 resultRank[index]++;
             }
         }
